(#1903) Divided CP module into subparts
Divided CP main module into
- main script
- metadata handling
- sql handling
- api/http requests
